+++ contracts/mocks/account/AccountMock.sol
@@ -6,7 +6,7 @@ import {PackedUserOperation} from "@openzeppelin/contracts/interfaces/draft-IERC
 import {ERC4337Utils} from "@openzeppelin/contracts/account/utils/draft-ERC4337Utils.sol";
 import {Account} from "../../account/Account.sol";
 
-abstract contract AccountBaseMock is Account {
+abstract contract AccountMock is Account {
     /// Validates a user operation with a boolean signature.
     function _rawSignatureValidation(
         bytes32 /* userOpHash */,